Finish stat window, put frontier into perf score.
Added/updated some detailed statistics to detail the current frontier coverage.
These involve the followings: visited-frontier ratio, proximity degree, average
proximity, and frontiers. These are more thoroughally explained within the
status_screen.txt file.

Added labels to the instrumentation code so that normal "fall-throughs" within
a label, or an unconditional jump to the label can be ignored and bypasses the
instrumentation. This helps to avoid tracking "false frontiers" that might
result when, for example, an unconditional jump jumps to a label marked as the
target of a conditional jump.

Fixed bugs in calculating frontier information. Uhh... yeah... that part of code
was pretty buggy. There were multiple nooby errors in all the bitwise
operations, but hopefully they are fixed.

Fixed bug when multiple labels alias to one location within code. This is
resolved by adding label aliases that "link" to the actual used label. We also
try to add suffixes to labels to avoid having duplicate labels within the
instrumented binary.

TODO:
  * Add colors for the frontier coverage statistics
  * Place stats in window into the respective stat file(s)
  * Implement filtering of input cases based on frontiers?